backwards compatible ways of reading local hazard
